+++ src/main/java/chess/Command.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess;
 
 import chess.domain.state.State;
 import java.util.Arrays;
