+++ src/test/java/chess/CommandTest.java
@@ -1,8 +1,9 @@
-package chess.domain;
+package chess;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
